smooth arrow
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@4388 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ AYDicomPrint/AYNSImageToDicom.m
@@ -234,7 +234,7 @@
 - (NSString*) generateUniqueFileName:(NSString*) destinationPath
 {
 	NSTimeInterval secs = [NSDate timeIntervalSinceReferenceDate];
-	NSString *filePath = [NSString stringWithFormat: @"%@/%ld", destinationPath, secs];
+	NSString *filePath = [NSString stringWithFormat: @"%@/%ld", destinationPath, (long) secs];
 	int index = 0;
 	BOOL isDir = YES;
